Q7
You have been provided with a database that contains information about employees and their salaries. Your task is to write a SQL query to retrieve the following information:

The average salary of all employees.The maximum salary of employees in each department.The total number of employees in each department.