Create Q12
You have been provided with a dataset that contains information about books and their authors. Your task is to create a Node.js application to perform some data manipulation operations on the dataset and answer the following questions:

What are the top 5 most popular book genres?Which author has written the most number of books?What is the average rating of books published by each author?Which book has the highest rating?